refactor: extract the join graph into its own translation unit
The optimize-join-order rule's header declared the whole join graph -- the
AttributePath alias, the JoinGraph structure and buildJoinGraph -- alongside
the rule function itself. Registering the rule therefore required seeing how
it works, and the graph could not be used or tested without pulling in the
rule header.

Move the graph and its construction into a unit of their own under a
subdirectory owned by the rule, leaving the rule header declaring the entry
point and nothing else. The filter classifier and the attribute-access helper
stay file-private, and the maintainer-mode trace stays with the rule that
calls it.

Every relocated body is byte-identical to its previous text; no behaviour
changes.
