+++ arangod/Aql/Optimizer/Rule/OptimizeJoinOrder/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(arango_aql PRIVATE
+  JoinGraph.cpp
+)
