+++ tests/Aql/OptimizeJoinOrderTest.cpp
@@ -23,7 +23,7 @@
 
 #include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionPlan.h"
-#include "Aql/Optimizer/Rule/OptimizeJoinOrder.h"
+#include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraph.h"
 #include "Aql/Query.h"
 #include "Aql/QueryString.h"
 #include "Aql/Variable.h"
